Move precise math in util
