+++ include/boost/geometry/strategy/cartesian/precise_area.hpp
@@ -18,7 +18,7 @@
 #include <boost/geometry/core/coordinate_dimension.hpp>
 #include <boost/geometry/strategy/area.hpp>
 #include <boost/geometry/util/select_most_precise.hpp>
-#include <boost/geometry/extensions/triangulation/strategies/cartesian/detail/precise_math.hpp>
+#include <boost/geometry/util/precise_math.hpp>
 
 namespace boost { namespace geometry
 {
